améliorations sauvegarde en continue et TV

+++ device/sav_impulsions.lua
@@ -5,28 +5,13 @@ dofile('./interface/interface.lua');
 -- Information : Numéro de Version, Nom, Interface
 function device.GetInformation()
 	return { 
-		version = 2.8, 
+		version = 2.9, 
 		name = 'Sauvegarde des impulsions', 
 		class = 'tools'
 	};
 end	
 
 function device.OnConfiguration(node)
-	function OnPath()
-		local dirDialog = wnd.CreateDirDialog({parent = dlgConfig, 
-			"Emplacement du fichier de sauvegarde", 
-			sav.directory
-			});
-		if dirDialog:ShowModal() == idButton.OK then
-			sav.directory = string.gsub(dirDialog:GetPath(), app.GetPathSeparator(), "/");
-			dlgConfig:GetWindowName('directory'):SetValue(sav.directory);
-		end	
-	end
-	function OnSaveConfig()
-		sav.directory = dlgConfig:GetWindowName('directory'):GetValue();
-		node:ChangeAttribute('directory', sav.directory);
-	end
-	
 	sav = sav or {};
 	sav.directory = node:GetAttribute('directory');
 	if sav.directory:len() < 2 then
@@ -42,6 +27,7 @@ function device.OnConfiguration(node)
 		label='Configuration du répertoire de sauvegarde', 
 		icon='./res/32x32_param.png'
 	});
+
 	dlgConfig:LoadTemplateXML({ 
 		xml = './device/sav_impulsions.xml', 	
 		node_name = 'root/panel', 			
@@ -49,15 +35,38 @@ function device.OnConfiguration(node)
 		node_value = 'config' 				
 	});
 
+	function OnPath()
+		local dirDialog = wnd.CreateDirDialog(
+			{ parent = dlgConfig },
+			"Emplacement du fichier de sauvegarde", 
+			sav.directory
+		);
+		if dirDialog:ShowModal() == idButton.OK then
+			sav.directory = string.gsub(dirDialog:GetPath(), app.GetPathSeparator(), "/");
+			dlgConfig:GetWindowName('directory'):SetValue(sav.directory);
+		end	
+		dirDialog:Delete();
+	end
+	
+	function OnSaveConfig()
+		sav.directory = dlgConfig:GetWindowName('directory'):GetValue();
+		node:ChangeAttribute('directory', sav.directory);
+	end
+
+	function OnClose()
+		dlgConfig:EndModal(idButton.CANCEL);
+	end
+
 	-- Toolbar Principale ...
 	local tb = dlgConfig:GetWindowName('tb');
 	local btnSave = tb:AddTool("Valider", "./res/32x32_save.png");
 	tb:AddStretchableSpace();
 	local btnClose = tb:AddTool("Fermer", "./res/32x32_close.png");
 	tb:Realize();
+	
 	dlgConfig:Bind(eventType.BUTTON, function(evt) OnPath() end, dlgConfig:GetWindowName('path'));
 	dlgConfig:Bind(eventType.MENU, function(evt) OnSaveConfig() end, btnSave);
-	tb:Bind(eventType.MENU, function(evt) dlgConfig:EndModal(idButton.CANCEL) end, btnClose);
+	tb:Bind(eventType.MENU, function(evt) OnClose() end, btnClose);
 	
 	dlgConfig:GetWindowName('directory'):SetValue(sav.directory);
 
@@ -99,22 +108,28 @@ function device.OnInit(params, node)
 	sosDb:Load();
 	tResultatChrono = sosDb:GetTable("Resultat_Chrono");
 	if tResultatChrono == nil then 
-		sosDb:Delete();
 		-- création de la base sav.db3 et création de la table Resultat_Chrono
+		sosDb:Delete();
+		
 		local base = sqlBase.Clone();
 		local tResultatChronoSki = base:GetTable("Resultat_Chrono");
-		tResultatChronoSki:RemoveAllRows();
-		tResultatChronoSki:Snapshot(sav.db3);
+		local cmd = "Select * From Resultat_Chrono Where Code_evenement = "..sav.code_evenement.." And Code_manche = "..sav.code_manche;
+		base:TableLoad(tResultatChronoSki, cmd);
+		local ignoreConstraint = false; -- On veut les primary key ...
+		tResultatChronoSki:Snapshot(sav.db3, ignoreConstraint);
+	
 		sosDb = sqlBase.ConnectSQLite(sav.db3);
 		sosDb:Load();
 		tResultatChrono = sosDb:GetTable("Resultat_Chrono");
 		base:Delete();	
-	else
-		sosDb:TableLoad(tResultatChrono, "Select * From Resultat_Chrono Order By Seq ASC");
-		seq = tResultatChrono:GetCellInt("Seq", tResultatChrono:GetNbRows() - 1,0);
 	end
-	assert(tResultatChrono ~= nil);
+
+	if tResultatChrono == nil or tResultatChrono:GetIndexColumn("Seq") < 0 then 
+		return;
+	end
 	
+	local seq = sosDb:SelectInt("Select Max(Seq) From Resultat_Chrono");
+
 	-- Creation Panel
 	panel = wnd.CreatePanel({
 		parent = app.GetAuiFrame(),
@@ -141,7 +156,7 @@ function device.OnInit(params, node)
 		show = true,
 		float = true, 
 		floating_position = {981, 25},
-		floating_size = {220, 110},
+		floating_size = {250, 90},
 		dockable = false
 	});
 	mgr:Update();
@@ -152,8 +167,7 @@ function device.OnInit(params, node)
 	local btn_recharger = tb:AddTool("Recharger les séquences", "./res/32x32_download.png");
 	tb:Realize();
 	
-	assert(panel:GetWindowName('sequence') ~= nil);
-	RefreshCompteur(0)
+	RefreshCompteur(seq)
 
 	-- Prise des Evenements (Bind)
 	app.BindNotify("<passage_insert>", OnNotifyPassageInsert);
@@ -166,11 +180,13 @@ function device.OnInit(params, node)
 	adv.Success("Ouverture du fichier "..sav.db3.." Ok ...");
 end
 
--- Fermeture
 function device.OnClose()
-	if panel ~= nil then
-		app.GetAuiManager():DeletePane(panel);
+	if sosDb ~= nil then
+		sosDb:Delete();
 	end
+
+	-- Appel OnClose Metatable
+	mt_device.OnClose();
 end
 
 function OnChangeRepertoire();
@@ -198,52 +214,63 @@ function OnRechargerSequence();
 		return;
 	end
 	--  rechercher le fichier .db3 des séquences à relire et le charger
-	local fileDialog = wnd.CreateFileDialog(panel,
+	local fileDialog = wnd.CreateFileDialog( 
+		{ parent = panel },
 		"Sélection du fichier de sauvegarde",
 		sav.directory, 
 		"",
 		"*.db3|*.db3",
 		fileDialogStyle.OPEN+fileDialogStyle.FD_FILE_MUST_EXIST
 	);
-	if fileDialog:ShowModal() == idButton.OK then
-		read_db3 = string.gsub(fileDialog:GetPath(), app.GetPathSeparator(), "/");
+	if fileDialog:ShowModal() ~= idButton.OK then
 		fileDialog:Delete();
-	else
-		fileDialog:Delete();
-		return false;
-	end
-	readDb = sqlBase.ConnectSQLite(read_db3);
-	readDb:Load();
-	if readDb:GetNbTables() == 0 then
-		panel:MessageBox("Le fichier sélectionné n'est pas un fichier valide !\n"..read_db3, "Erreur", msgBoxStyle.OK+msgBoxStyle.ICON_ERROR);
 		return false;
 	end
 	
+	if sosDb ~= nil then
+		sosDb:Delete();
+		sosDb = nil;
+	end
+
+	read_db3 = string.gsub(fileDialog:GetPath(), app.GetPathSeparator(), "/");
+	fileDialog:Delete();
+	sosDb = sqlBase.ConnectSQLite(read_db3);
+	sosDb:Load();
+
+	if sosDb:GetTable('Resultat_Chrono') == nil then
+		panel:MessageBox("Le fichier sélectionné n'est pas un fichier valide !\n"..read_db3, "Erreur", msgBoxStyle.OK+msgBoxStyle.ICON_ERROR);
+		sosDb:Delete();
+		sosDb = nil;
+		return false;
+	end
+
 	base = sqlBase.Clone();
 	cmd = "Delete From Resultat_Chrono Where Code_evenement = "..sav.code_evenement.." And Code_manche = "..sav.code_manche;
 	base:Query(cmd);
 	
 	-- Chargement Table DB3
-	readDb:TableLoad(tResultatChrono, "Select * From Resultat_Chrono Order By Seq ASC");
+	local tResultatChrono = sosDb:GetTable('Resultat_Chrono');
+	sosDb:TableLoad(tResultatChrono, "Select * From Resultat_Chrono Order By Seq ASC");
 
 	tResultatChronoSki = base:GetTable("Resultat_Chrono");
+	rResultatChronoSki = tResultatChronoSki:GetRecord();
+	
 	for row = 0, tResultatChrono:GetNbRows()-1 do
-		tResultatChronoSki:AddRow();
-		tResultatChronoSki:SetCell('Code_evenement',row, sav.code_evenement);
-		tResultatChronoSki:SetCell('Code_manche',row, sav.code_manche);
-		tResultatChronoSki:SetCell('Seq', row, tResultatChrono:GetCellInt("Seq",row));
-		tResultatChronoSki:SetCell('Origine', row, tResultatChrono:GetCell("Origine",row));
-		tResultatChronoSki:SetCell('Heure', row, tResultatChrono:GetCellInt("Heure",row));
-		tResultatChronoSki:SetCell('Dossard', row, tResultatChrono:GetCell("Dossard",row));
-		tResultatChronoSki:SetCell('Dossard_anc', row, tResultatChrono:GetCell("Dossard_anc",row));
-		tResultatChronoSki:SetCell('Id', row, tResultatChrono:GetCellInt("Id",row));
-		tResultatChronoSki:SetCell('Device', row, tResultatChrono:GetCell("Device",row));
-		base:TableInsert(tResultatChronoSki, row);
+		rResultatChronoSki:Set('Code_evenement', sav.code_evenement);
+		rResultatChronoSki:Set('Code_manche', sav.code_manche);
+		rResultatChronoSki:Set('Seq', tResultatChrono:GetCellInt("Seq",row));
+		rResultatChronoSki:Set('Origine', tResultatChrono:GetCell("Origine",row));
+		rResultatChronoSki:Set('Heure', tResultatChrono:GetCellInt("Heure",row));
+		rResultatChronoSki:Set('Dossard', tResultatChrono:GetCell("Dossard",row));
+		rResultatChronoSki:Set('Dossard_anc', tResultatChrono:GetCell("Dossard_anc",row));
+		rResultatChronoSki:Set('Id', tResultatChrono:GetCellInt("Id",row));
+		rResultatChronoSki:Set('Device', tResultatChrono:GetCell("Device",row));
+		base:TableInsert(tResultatChronoSki, -1);
 	end
-	
 	base:Delete();
-	readDb:Delete();
-	
+
+	RefreshCompteur(sosDb:SelectInt("Select Max(Seq) From Resultat_Chrono"));
+
 	-- ReChargement des Passages dans la fenêtre de chronométrage
 	-- à faire un recalcul des temps nets  partir des séquences quelle est la notification à envoyer
 	app.SendNotify('<run_reload>'); 
@@ -254,12 +281,15 @@ function OnRunErase(key, params);
 	if panel:MessageBox("Voulez-vous également effacer les impulsions \ndéjà sauvegardées pour la manche "..sav.code_manche.." ?", "Confirmation de l'effacement de la sauvegerde", msgBoxStyle.YES_NO+msgBoxStyle.ICON_INFORMATION) ~= msgBoxStyle.YES then
 		return;
 	end
-	tResultat_Chrono = sosDb:GetTable('Resultat_Chrono');
-	if tResultat_Chrono ~= nil then 
-		cmd = "Delete From Resultat_Chrono Where Code_manche = "..sav.code_manche;
-		sosDb:Query(cmd);
+
+	if sosDb ~= nil then
+		tResultat_Chrono = sosDb:GetTable('Resultat_Chrono');
+		if tResultat_Chrono ~= nil then 
+			cmd = "Delete From Resultat_Chrono Where Code_manche = "..sav.code_manche;
+			sosDb:Query(cmd);
+		end
+		RefreshCompteur(0)
 	end
-	RefreshCompteur(0)
 end
 
 function OnNotifyPassageInsert(key, params)
@@ -273,7 +303,7 @@ function OnNotifyPassageInsert(key, params)
 	local bib = params.bib;
 	local id = tonumber(params.passage);
 	local device = params.device;
-	if id == -1 then	-- ça mange pas de pain ....
+	if id == -1 then	-- ça mange pas de pain (compatibilité FFSSSKI ...)
 		origine = "A";
 	elseif id == 0 then
 		origine = "D";
@@ -285,8 +315,11 @@ function OnNotifyPassageInsert(key, params)
 	else
 		cmd = "Insert Into Resultat_Chrono (Code_evenement, Code_manche, Seq, Origine, Heure, Dossard, Dossard_anc, Id, Device) Values("..sav.code_evenement..","..sav.code_manche..","..seq..',"'..origine..'",'..heure..', NULL, NULL, '..id..',"'..device..'")';
 	end
-	sosDb:Query(cmd);
-	RefreshCompteur(seq)
+	
+	if sosDb ~= nil then
+		sosDb:Query(cmd);
+		RefreshCompteur(seq)
+	end
 end
 
 function OnNotifyPassageUpdate(key, params)
@@ -311,14 +344,15 @@ function OnNotifyPassageUpdate(key, params)
 				" Where Code_manche = "..sav.code_manche.." And Seq = "..seq;
 		end
 	end
+	
 	-- update de l'impulsion dans la base
-	sosDb:Query(cmd);
+	if sosDb ~= nil then
+		sosDb:Query(cmd);
+	end
 end
 
 function RefreshCompteur(seq)
-	if seq ~= nil then
-		panel:GetWindowName('sequence'):SetLabel('sequence : '..tostring(seq));
-	else
-		panel:GetWindowName('sequence'):SetLabel('sequence : ');
+	if seq ~= nil and panel:GetWindowName('sequence') ~= nil then
+		panel:GetWindowName('sequence'):SetLabel('Seq : '..tostring(seq));
 	end
 end
